requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
distlib==0.4.0
distro==1.9.0
filelock==3.24.3
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.11
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.18
langchain-openai==1.1.11
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.1
langchain-upstage==0.7.7
langgraph==1.1.0
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.16
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.18.0
numpy==2.4.3
openai==2.26.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
pillow==12.1.1
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.9.2
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pypdf==6.8.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-discovery==1.1.0
python-docx==1.2.0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
screeninfo==0.8.1
simsimd==6.5.16
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
streamlit==1.55.0
streamlit-js-eval==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.20.3
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
virtualenv==21.1.0
watchdog==6.0.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
